requirements.txt
absl-py==1.4.0
appdirs==1.4.4
argcomplete==3.1.1
astral==3.2
astunparse==1.6.3
audioread==3.0.0
backports.zoneinfo==0.2.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==41.0.2
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
ephem==4.1.4
filelock==3.12.2
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
geographiclib==2.0
geopy==2.3.0
google-auth==2.21.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonify==0.5
keras==2.11.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
librosa==0.10.0.post2
llvmlite==0.39.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
mpmath==1.3.0
msgpack==1.0.5
Naked==0.1.32
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pipx==1.2.0
platformdirs==3.8.1
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
shellescape==3.8.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.7.1
urllib3==1.26.16
userpath==1.8.0
virtualenv==20.23.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
